requirements.txt
argcomplete==3.5.3
asgiref==3.8.1
astroid==3.2.4
babel==2.16.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.10.0
blivet-gui==2.6.0
boto3==1.35.93
botocore==1.35.93
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==3.3
click-plugins==1.1.1
contourpy==1.3.0
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
defusedxml==0.8.0rc2
dill==0.3.8
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-stubs==5.0.0
django-stubs-ext==5.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dnf==4.22.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.14.0
fonttools==4.53.1
fros==1.1
gopen==0.6.6
greenlet==3.0.3
humanize==3.13.1
idna==3.7
iso639==0.1.4
isort==5.13.2
jmespath==1.0.1
kiwisolver==1.4.7
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.4
libvirt-python==10.1.0
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
MarkupSafe==2.1.3
matplotlib==3.9.2
mccabe==0.7.0
meson==1.4.1
msgpack==1.0.6
mutagen==1.47.0
mysqlclient==2.2.7
nftables==0.1
numpy==2.0.1
oauthlib==3.2.2
olefile==0.47
packaging==23.2
Paste==3.7.1
penv==1.2
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.2
ply==3.11
productmd==1.43
progressbar2==3.53.2
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==2.0.4
pyenchant==3.2.2
pygit2==1.14.0
PyGObject==3.48.2
PyJWT==2.10.1
pykickstart==3.52
pylint==3.2.6
pynvim==0.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.4
pyparted==3.13.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-magic==0.4.27
python-meh==0.51
python-pam==2.0.2
python-utils==3.7.0
python3-openid==3.2.0
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
reportlab==4.2.4
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
rpm==4.19.1.1
rpmautospec==0.7.3
rpmautospec-core==0.1.5
s3transfer==0.10.4
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sos==4.7.2
soupsieve==2.5
sqlparse==0.5.0
systemd-python==235
Tempita==0.5.2
tomlkit==0.13.2
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
urllib3==1.26.20
vboxapi==1
virtualenv==20.26.2
websockets==12.0
xkbregistry==0.3
yt-dlp==2025.1.15
